[a11y] Disable buttons and scrubber when modal dialogs are open
Modal dialogs are implemented using MaterialUI Dialog class.  Dialog elements block navigation outside the modal using JS and sets the rest of the page aria-hidden but does not do anyother changes to the underlying page.  This leaves accessibility checkers like SiteImprove to detect navigable aria-hidden elements.  This feels a bit like a false positive but this PR does the work to disable those interactive elements when modal dialogs are open.